Added stats updates

See merge request cs309/fall2019/VB_1!35

+++ BackEnd/CydeNSeek/src/main/java/application/model/Stats.java
@@ -30,7 +30,7 @@ public class Stats {
 	private Integer gwseeker;
 
 	@Column
-	private Integer totdistance;
+	private Double totdistance;
 
 	@Column
 	private Integer tottime;
@@ -67,11 +67,11 @@ public class Stats {
 		this.gwseeker = gwseeker;
 	}
 
-	public Integer getTotDistance() {
+	public Double getTotDistance() {
 		return totdistance;
 	}
 
-	public void setTotDistance(Integer totdistance) {
+	public void setTotDistance(Double totdistance) {
 		this.totdistance = totdistance;
 	}
